Enable TextureMapper and fix compilation
For now software AC mode (which is not supported in upstream) is removed.
It can be restored separately later.

Primary goal of this patch is to prevent liking errors in code which invokes
GraphicsLayer::create(). Enabling AC in OpenGL mode is not yet possible
because bulding with ENABLE_OPENGL is not supported.

Change-Id: I5000c2f050bd106395a10531e7a59c62f2efa01d