SVG fragment identifier is not respected if it is a part of an HTTP URL
This change is necessary for good user experience in UI of remote
inspector, in particular to make triangles in DOM tree, which allow to
collapse or expand nodes, visible.

Some SVG images rely on fragment to be the input parameter which changes
their shape. However, caching subsystem strips out fragment to use the
same cached object for URL with any fragment. This is logical, but
unfortunate side effect is that this fragment is not then passed when SVG
is rendered.

This patch makes all image resources to keep their fragment as cache key.
It leads to negative side effect: images with fragments in URL are cached
as much many times as many different fragments are used, and not only
SVG images but other formats too.

Proper fix is available in master but it's too hard to backport. However,
I think that regression is probably less harmful then original bug.

Task-number: 818
Change-Id: I19ef20c74dd07e54911b1640f7fa0fddc2fbcd4c