[conan] Set CC and CXX variables when building with pre-existing profile
CMake Ninja generator requires CC and CXX variables to do the right choice,
and Conan doesn't set them automatically. When we are generating profile,
CC and CXX are set to match chosen compiler. Now we also set them when
--profile option is used.

Note that if profile defines CC and/or CXX in its [env], profile values
take priority.

Task-number: 1033