[WebRTC][GStreamer] fast/mediastream/resize-trim.html is failing https://bugs.webkit.org/show_bug.cgi?id=192888
Patch by Philippe Normand <pnormand@igalia.com> on 2021-05-04
Reviewed by Xabier Rodriguez-Calvar.

Source/WebCore:

The test was failing because the GStreamer mock video source was producing garbled frames
and the pixel test in the canvas was failing to detect gray. The frame was garbled because
the test asked "exact" video width that was inferior to the capture video width and we were
not taking this into account when creating a MediaSample from the video frame.

MediaSampleGStreamer::createImageSample() can now optionally resize the video frame
according to the destinationSize, if it was specified.

* html/HTMLCanvasElement.cpp:
(WebCore::HTMLCanvasElement::toMediaSample):
* platform/graphics/gstreamer/MediaSampleGStreamer.cpp:
(WebCore::MediaSampleGStreamer::createImageSample):
* platform/graphics/gstreamer/MediaSampleGStreamer.h:
(WebCore::MediaSampleGStreamer::createImageSample):
* platform/mediastream/gstreamer/MockRealtimeVideoSourceGStreamer.cpp:
(WebCore::MockRealtimeVideoSourceGStreamer::updateSampleBuffer):
* platform/mock/MockRealtimeVideoSource.h:

LayoutTests:

* platform/glib/TestExpectations: Unflag test.
* resources/platform-helper.js: Remove EFL cruft.
(videoCanvasPixelComparisonTolerance):
(isEfl): Deleted.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@276953 268f45cc-cd09-0410-ab3c-d52691b4dbfc