[Cocoa] Cache the value of MediaPlayerPrivateAVFoundationObjC::assetStatus() https://bugs.webkit.org/show_bug.cgi?id=225262
Reviewed by Eric Carlson.

Calling into assetStatus() requires querying the load state and value of nine different
AVAsset properties, which adds up if assetStatus() is called often. Once the assetStatus()
value reaches the Loaded state, cache that property, and that of the AVAsset.playable
property, so subsequent queries are just an ivar lookup.

* platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.h:
* platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.mm:
(WebCore::MediaPlayerPrivateAVFoundationObjC::assetStatus const):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@276981 268f45cc-cd09-0410-ab3c-d52691b4dbfc