[macOS] Use system colors for ActiveButtonText https://bugs.webkit.org/show_bug.cgi?id=225375 <rdar://problem/75863443>
Reviewed by Tim Horton.

Currently, the ActiveButtonText color keyword returns a hardcoded value
on macOS. This behavior is problematic since AppKit can make changes to
the button style, leaving us with illegible text when a button is
pressed.

To fix, return a system color for ActiveButtonText, so that our pressed
buttons are consistent with the rest of the system.

* rendering/RenderThemeMac.mm:
(WebCore::activeButtonTextColor):

AppKit determines the pressed (active) text color using the cell's
interiorBackgroundStyle. There is not a single system color exposed for
the pressed text color, so it must be determined dynamically.

Unfortunately, this requires us to create an NSButtonCell to determine
the appropriate color. However, this logic is relatively inexpensive,
since we cache exposed system colors.

(WebCore::RenderThemeMac::systemColor const):

Remove the hardcoded value.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277067 268f45cc-cd09-0410-ab3c-d52691b4dbfc