[Metal ANGLE] Only clear dirty state bits after all state has been successfully set https://bugs.webkit.org/show_bug.cgi?id=225468 <rdar://77066752>
When running in release, if any ANGLE calls fail, we eventually return to this part of the code.
We can be left with an incomplete state, where we have an incomplete program or attachment,
but a 'clean' state. This can lead to errors such as setting nil render pipeline states, leading to
crashes.

To mitigate this, hold off on setting state bits until we've completly set up a program.

Reviewed by Dean Jackson.

* src/libANGLE/renderer/metal/ContextMtl.mm:
(rx::ContextMtl::setupDraw):
* src/libANGLE/renderer/metal/ProgramMtl.mm:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277143 268f45cc-cd09-0410-ab3c-d52691b4dbfc