Crash in InsertTextCommand::positionInsideTextNode https://bugs.webkit.org/show_bug.cgi?id=223753
Patch by Frédéric Wang <fwang@igalia.com> on 2021-05-07
Reviewed by Ryosuke Niwa.

Source/WebCore:

Test: editing/deleting/selection-on-empty-table-row.html

This is a follow-up of bug 213514, where removePreviouslySelectedEmptyTableRows was
modified to ensure the selection is properly updated after deleting the last row. This
patch does the same for other rows.

* editing/DeleteSelectionCommand.cpp:
(WebCore::DeleteSelectionCommand::removePreviouslySelectedEmptyTableRows): Update comment
about calling raw removeNode and move it to the top of the function. Replace all the calls
to CompositeEditCommand::removeNode with CompositeEditCommand::removeNodeUpdatingStates so
that the selection is adjusted.

LayoutTests:

Add regression test.

* editing/deleting/selection-on-empty-table-row-expected.txt: Added.
* editing/deleting/selection-on-empty-table-row.html: Added.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277163 268f45cc-cd09-0410-ab3c-d52691b4dbfc