[iOS] fix inconsistency around the meaning of `rate` and `defaultPlaybackRate` between WebKit and AVKit https://bugs.webkit.org/show_bug.cgi?id=225532 <rdar://problem/77629276>
Reviewed by Eric Carlson.

In AVKit, the `defaultPlaybackRate` is used when playback starts, such as resuming after
pausing. In WebKit, however, `defaultPlaybackRate` is only used when first loading and after
ending scanning, with the `playbackRate` being used in all other cases, including when
resuming after pausing. As such, WebKit should return the `playbackRate` instead of the
`defaultPlaybackRate` in these cases when communicating with AVKit.

* platform/ios/WebAVPlayerController.mm:
(-[WebAVPlayerController setRate:]):
Also `setDefaultPlaybackRate:` if not paused (i.e. `rate != 0`) since WebKit's definition of
`playbackRate` does not change when paused.

* platform/ios/PlaybackSessionInterfaceAVKit.mm:
(WebCore::PlaybackSessionInterfaceAVKit::rateChanged):
Make sure to `setDefaultPlaybackRate:` before `setRate:` as the latter can affect the former.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277203 268f45cc-cd09-0410-ab3c-d52691b4dbfc