[JSC] Error#cause must recognize explicit undefined https://bugs.webkit.org/show_bug.cgi?id=225535
Reviewed by Alexey Shvayka.

JSTests:

* test262/config.yaml:
Re-enable tests for this feature; they were all failing due to this quirk.

Source/JavaScriptCore:

Error#cause is specified such that `new Error(message, {})` and `new Error(message, { cause: undefined })`
are not the same -- namely, the latter should create a property descriptor with an undefined `value`.

This would seem absurd, but the reason is because the `cause` field is meant to store a thrown object,
and `throw undefined;` is valid code.

In aligning our implementation with the spec, this patch also consolidates the relevant logic in one place
(ErrorInstance::finishCreation) to minimize confusion.

* runtime/AggregateError.cpp:
(JSC::createAggregateError):
* runtime/ErrorInstance.cpp:
(JSC::ErrorInstance::create):
(JSC::ErrorInstance::finishCreation):
* runtime/ErrorInstance.h:
(JSC::ErrorInstance::create):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277221 268f45cc-cd09-0410-ab3c-d52691b4dbfc