[ macOS Release wk2 ] http/tests/cache-storage/cache-records-persistency.https.html is flaky crashing https://bugs.webkit.org/show_bug.cgi?id=225522 <rdar://problem/77680019>
Reviewed by Darin Adler.

Source/WTF:

Pass ec parameter to std::filesystem::directory_iterator() so that it doesn't throw in case of the
path does not exist. Normally, the function would return early earlier if the path does not exist
or isn't a directory. However, in the context of the test, this function is getting called on a
background thread and another thread is deleting the directory in parallel.

* wtf/FileSystem.cpp:
(WTF::FileSystemImpl::deleteEmptyDirectory):

LayoutTests:

Restore previous expectation for the test, now that it is no longer flaky crashing.

* platform/mac-wk2/TestExpectations:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277315 268f45cc-cd09-0410-ab3c-d52691b4dbfc