[GStreamer] media/track/in-band/track-in-band-srt-mkv-kind.html is a flaky crash https://bugs.webkit.org/show_bug.cgi?id=225697
Source/WebCore:

This is a tentative fix for the flaky crash (I can't reproduce it locally after 10000
repetitions). It makes sure that mediaPlayerPrivate and any parameter passed to
handleTextSample() can either survive or detected to be invalid when the lambda is ran in
the main thread.

Reviewed by Alicia Boya Garcia.

Covered by existing tests.

* platform/graphics/gstreamer/TextSinkGStreamer.cpp:
(webkitTextSinkHandleSample): Protect mediaPlayerPrivate and streadId and check if the mediaPlayerPrivate WeakRef is still alive.

LayoutTests:

Reviewed by Alicia Boya Garcia.

* platform/glib/TestExpectations: Unskipped test.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@277432 268f45cc-cd09-0410-ab3c-d52691b4dbfc